requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiohttp==3.9.1
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
numpy==1.26.2 